revert TestUpdateFeaturedSnaps location
